+++ lib/rstat.rb
@@ -1,5 +1,6 @@
-require "rstat/version"
-require "rstat/core_ext"
+Dir["#{File.dirname(__FILE__)}/rstat/*.rb"].sort.each do |path|
+  require "rstat/#{File.basename(path, '.rb')}"
+end
 
 module Rstat
 end