+++ lib/rstat/core_ext/array.rb
@@ -1,3 +1,3 @@
 Dir["#{File.dirname(__FILE__)}/array/*.rb"].sort.each do |path|
-  require "rstat/core_ext/array/#{File.basename(path, ".rb")}"
+  require "rstat/core_ext/array/#{File.basename(path, '.rb')}"
 end