+++ spec/spec_helper.rb
@@ -1,2 +1,2 @@
 # encoding: utf-8
-require "rstat"
+require 'rstat'